requirements.txt
blinker
click
Flask
itsdangerous
Jinja2
MarkupSafe
packaging
pillow
qrcode
Werkzeug
